Refactor listings search
