--- components/listings/Search/fields/RangeComponent/styles.js
@@ -1,20 +1,0 @@
-import {StyleSheet} from 'react-native'
-
-import {margin} from '@/assets/styles'
-
-export default StyleSheet.create({
-  container: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center'
-  },
-  field: {
-    flex: 0.5
-  },
-  divider: {
-    fontWeight: '600',
-    textAlign: 'center',
-    ...margin(null, 25)
-  }
-})
