+++ containers/views/Listings/Search/index.js
@@ -1,16 +1,14 @@
 import {StackNavigator} from 'react-navigation'
 
 import Header from './Header'
-import * as menu from './Menu'
-import fields from './fields'
+import * as filter from './Filter'
 
 export const screen = StackNavigator(
   {
-    menu,
-    ...fields
+    filter
   },
   {
-    initialRouteName: 'menu',
+    initialRouteName: 'filter',
     initialRouteParams: {},
     headerMode: 'float',
     navigationOptions: {
